removed old ui code
